requirements.txt
aws-cdk.cdk==0.33.0
aws-cdk.aws_kinesisfirehose==0.33.0
aws-cdk.aws_s3==0.33.0
aws-cdk.aws_iam==0.33.0
aws-cdk.aws_logs==0.33.0
aws-cdk.aws_ec2==0.33.0
aws-cdk.aws_ecs==0.33.0
aws-cdk.aws_ssm==0.33.0
aws-cdk.aws_glue==0.33.0
python-twitter==3.5
boto3==1.9.159
